Refactor: Moves helpers out of main
